chore: clean up .gitignore and update engine/card data

+++ engine_rust_src/src/core/logic/interpreter/handlers/movement_discard_prompt.rs
@@ -104,7 +104,7 @@ pub fn prepare_discard_prompt(
                     O_MOVE_TO_DISCARD,
                     s,
                     choice_type,
-                    filter_attr_with_mask,
+                    filter_attr_with_mask as u64,
                     v as i16,
                 ),
                 HandlerResult::Suspend
